backend/requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
google-cloud-speech
google-cloud-spanner
